Create README.md
An example project for demonstrating automation on a shopping website. 
The project is built in Java and Selenium on the TestNG infrastructure and demonstrates the use of important annotations and the listeners interfaces. The project is built according to the Page Object design pattern and contains a class for each page on the site. 
In addition, there is an abstract class for all reusable code, and a base class in which the appropriate driver initialization is defined and other central actions that are performed several times during the test. The tests take data from an external Json file.